refactor(web): extract shared in-memory runQuery engine from fake-fetcher
